feat(autocomplete): add functionality to listen to input value changes
Previously, the Autocomplete component did not have the ability to listen to changes in the input value. This commit adds the `onInputChange` prop to the Autocomplete component, allowing developers to capture and respond to input value changes. The `onInputChange` callback function will receive the `event` object containing the updated input value. This enhancement enables users to perform actions based on the input value, such as making external API calls or updating other components. This change improves the flexibility and interactivity of the Autocomplete component.